Problems
While I was able to start working on a graphing class that is nearly finished, I few bugs were idntified. The primary bug which I suspect is causing the other bugs is the simulation being dependent on the clock. Due to this, the system can gain energy when low clock cycles cause poor simulation accuracy. Similarly, a system can lose energy through the same mechanism. Making the framerate more stable will be very important if this method is going to be continued, otherwise, another timing method will be necissary.