Added Springs!
I added a spring point class. For it to have general use I would have to update it in a few ways, but it can be set up with a "mesh" of the shape you want. I need to perfect it, but the ball sim shows how to generate a ball mesh. Everything seems to work except for rotational inertia seems to disapear... I know this has to do with how I damp, but I couldn't immediately figure it out.